Typescript

+++ src/components/Appointment/index.tsx
@@ -5,9 +5,11 @@ import Empty from './Empty';
 import Form from './Form';
 import Confirm from './Confirm';
 import Error from './Error';
-import useVisualMode from 'hooks/useVisualMode';
-import 'components/Appointment/styles.scss';
+import useVisualMode from '../../hooks/useVisualMode';
+import './styles.scss';
 import Status from './Status';
+import { AppointmentProps } from '../../helpers/propTypes';
+import { Interviewer } from '../../helpers/stateTypes';
 
 const EMPTY = 'EMPTY';
 const SHOW = 'SHOW';
@@ -19,30 +21,41 @@ const EDIT = 'EDIT';
 const ERROR_SAVE = 'ERROR_SAVE';
 const ERROR_DELETE = 'ERROR_DELETE';
 
-export default (props) => {
-  const { time, interview, interviewers, bookInterview, id, deleteInterview } =
-    props;
+export default ({
+  time,
+  interview,
+  interviewers,
+  bookInterview,
+  id,
+  deleteInterview,
+}: AppointmentProps) => {
   const { mode, transition, back } = useVisualMode(interview ? SHOW : EMPTY);
 
-  const save = (name, interviewer) => {
-    const interview = {
-      student: name,
-      interviewer,
-    };
-    transition(SAVING);
-    bookInterview(id, interview)
-      .then(() => transition(SHOW))
-      .catch((error) => transition(ERROR_SAVE, true));
+  const appointmentId: number | undefined = id !== undefined ? id : undefined;
+
+  const save = (name: string, interviewer: Interviewer) => {
+    if (bookInterview) {
+      const interview = {
+        student: name,
+        interviewer,
+      };
+      transition(SAVING);
+      bookInterview(id, interview)
+        .then(() => transition(SHOW))
+        .catch((error: unknown) => transition(ERROR_SAVE, true));
+    }
   };
 
   const deleteAppt = () => transition(CONFIRM);
   const editAppt = () => transition(EDIT);
 
-  const confirmDeleteAppt = (id) => {
-    transition(DELETE, true);
-    deleteInterview(id)
-      .then(() => transition(EMPTY))
-      .catch((error) => transition(ERROR_DELETE, true));
+  const confirmDeleteAppt = (id: number) => {
+    if (deleteInterview) {
+      transition(DELETE, true);
+      deleteInterview(id)
+        .then(() => transition(EMPTY))
+        .catch((error: unknown) => transition(ERROR_DELETE, true));
+    }
   };
 
   useEffect(() => {
@@ -55,14 +68,24 @@ export default (props) => {
   }, [interview, mode, transition]);
 
   return (
-    <article className="appointment">
+    <article className='appointment'>
       <Header time={time} />
       {mode === EMPTY && <Empty onAdd={() => transition(CREATE)} />}
       {mode === SHOW && interview && (
-        <Show onDelete={deleteAppt} {...interview} id={id} onEdit={editAppt} />
+        <Show
+          onDelete={deleteAppt}
+          {...interview}
+          id={appointmentId}
+          onEdit={editAppt}
+        />
       )}
       {mode === CREATE && (
-        <Form interviewers={interviewers} onCancel={back} onSave={save} />
+        <Form
+          interviewers={interviewers}
+          onCancel={back}
+          onSave={save}
+          student={interview ? interview.student : ''}
+        />
       )}
       {mode === SAVING && <Status message={'Saving...'} />}
       {mode === CONFIRM && (
@@ -79,8 +102,8 @@ export default (props) => {
           interviewers={interviewers}
           onCancel={back}
           onSave={save}
-          student={interview.student}
-          interviewer={interview.interviewer.id}
+          student={interview!.student}
+          interviewer={interview!.interviewer.id}
         />
       )}
       {mode === ERROR_SAVE && (
